requirements.txt
flask
python-dotenv
flask-sqlalchemy
flask-migrate
email-validator